+++ package/src/tools_pex.rs
@@ -55,7 +55,7 @@ pub(crate) fn build_tools_pex(
                     "--style",
                     "universal",
                     "--pip-version",
-                    "22.3",
+                    "23.3.2",
                     "--resolver-version",
                     "pip-2020-resolver",
                     "--no-build",