+++ tools/src/scie_pikesquares/install_pikesquares.py
@@ -17,57 +17,11 @@ from tinydb import TinyDB, Query
 from packaging.version import Version
 from platformdirs import user_data_dir
 
-from scie_pikesquares.log import debug, fatal, info, init_logging
+from scie_pikesquares.log import debug, fatal, info, init_logging, warn
 from scie_pikesquares.ptex import Ptex
 
 log = logging.getLogger(__name__)
 
-
-def venv_pip_install(venv_dir: Path, *args: str, find_links: str | None) -> None:
-    subprocess.run(
-        args=[
-            str(venv_dir / "bin" / "python"),
-            "-sE",
-            "-m",
-            "pip",
-            # This internal 1-use pip need not nag the user about its up-to-date-ness.
-            "--disable-pip-version-check",
-            "--no-python-version-warning",
-            "--log",
-            str(venv_dir / "pikesquares-install.log"),
-            "install",
-            "--quiet",
-            *(("--find-links", find_links) if find_links else ()),
-            *args,
-        ],
-        check=True,
-    )
-
-def install_pants_from_req(
-    venv_dir: Path, prompt: str, pants_requirements: Iterable[str], find_links: str | None
-) -> None:
-    subprocess.run(
-        args=[
-            sys.executable,
-            "-m",
-            "venv",
-            "--clear",
-            "--prompt",
-            prompt,
-            str(venv_dir),
-        ],
-        check=True,
-    )
-
-    # Pin Pip to 22.3.1 (currently latest). The key semantic that should be preserved by the Pip
-    # we use is that --find-links are used as a fallback only and PyPI is preferred. This saves us
-    # money by avoiding fetching wheels from our S3 bucket at https://binaries.pantsbuild.org unless
-    # absolutely needed.
-    #
-    # Also, we don't advance setuptools past 58 which drops support for the `setup` kwarg `use_2to3`
-    # which Pants 1.x sdist dependencies (pystache) use.
-    venv_pip_install(venv_dir, "-U", "pip==22.3.1", "setuptools<58", "wheel", find_links=find_links)
-    venv_pip_install(venv_dir, "--progress-bar", "off", *pants_requirements, find_links=find_links)
 
 def install_pikesquares_from_pex(
     venv_dir: Path,
@@ -101,22 +55,86 @@ def install_pikesquares_from_pex(
             check=True,
         )
 
+def install_pikesquares_localdev(
+    venv_dir: Path,
+    pikesquares_localdev_dir: str,
+) -> None:
+
+    subprocess.run(
+        args=[
+            sys.executable,
+            "-m",
+            "venv",
+            "--clear",
+            str(venv_dir),
+        ],
+        check=True,
+    )
+    #venv_pip_install(venv_dir, "-U", "pip==22.3.1", "setuptools<58", "wheel", find_links=find_links)
+    #venv_pip_install(venv_dir, "--progress-bar", "off", *pants_requirements, find_links=find_links)
+
+    subprocess.run(
+        args=[
+            str(venv_dir / "bin" / "python"),
+            "-sE",
+            "-m",
+            "pip",
+            "--disable-pip-version-check",
+            "--no-python-version-warning",
+            "--log",
+            str(venv_dir / "pikesquares-install.log"),
+            "install",
+            #"--quiet",
+            "--requirement",
+            str(Path(pikesquares_localdev_dir) / "requirements.txt"),
+        ],
+        check=True,
+    )
+
+    subprocess.run(
+        args=[
+            str(venv_dir / "bin" / "python"),
+            "-sE",
+            "-m",
+            "pip",
+            "--disable-pip-version-check",
+            "--no-python-version-warning",
+            "--log",
+            str(venv_dir / "pikesquares-install.log"),
+            "install",
+            #"--quiet",
+            "--editable",
+            pikesquares_localdev_dir,
+        ],
+        check=True,
+    )
+
+    pywsgi_wheel = "pikesquares_pyuwsgi-2.0.24.post0-cp312-cp312-manylinux_2_17_x86_64.manylinux2014_x86_64.whl"
+    pyuwsgi = f"https://pypi.vc.eloquentbits.com/packages/{pywsgi_wheel}"
+
+    subprocess.run(
+        args=[
+            str(venv_dir / "bin" / "python"),
+            "-sE",
+            "-m",
+            "pip",
+            "--disable-pip-version-check",
+            "--no-python-version-warning",
+            "--log",
+            str(venv_dir / "pikesquares-install.log"),
+            "install",
+            #"--quiet",
+            pyuwsgi,
+        ],
+        check=True,
+    )
+
 
 def main() -> NoReturn:
     parser = ArgumentParser()
     get_ptex = Ptex.add_options(parser)
     parser.add_argument(
         "--pikesquares-version", type=Version, required=True, help="The PikeSquares version to install"
-    )
-    parser.add_argument(
-        "--find-links",
-        type=str,
-        help="The find links repo pointing to PikeSquares pre-built wheels for the given PikeSquares version",
-    )
-    parser.add_argument(
-        "--pikesquares-bootstrap-urls",
-        type=str,
-        help="The path to the JSON file containing alternate URLs for downloaded artifacts.",
     )
     parser.add_argument("--debug", type=bool, help="Install with debug capabilities.")
     parser.add_argument("base_dir", nargs=1, help="The base directory to create PikeSquares venvs in.")
@@ -134,24 +152,38 @@ def main() -> NoReturn:
     venvs_dir = base_dir / "venvs"
 
     version = options.pikesquares_version
-    python_version = ".".join(map(str, sys.version_info[:3]))
-    info(f"Bootstrapping PikeSquares {version}")
-    debug(f"PikeSquares itself is using: {sys.implementation.name} {python_version}")
+    debug(f"lift.bindings.install: Bootstrapping PikeSquares {version}")
+    #python_version = ".".join(map(str, sys.version_info[:3]))
+    #debug(f"PikeSquares itself is using: {sys.implementation.name} {python_version}")
 
-    pikesquares_requirements = [f"pikesquares=={version}"]
     venv_dir = venvs_dir / str(version)
-    prompt = f"PikeSquares {version}"
 
-    info(
-        f"Installing {' '.join(pikesquares_requirements)} into a virtual environment at {venv_dir}"
-    )
-    install_pikesquares_from_pex(
-        venv_dir=venv_dir,
-        prompt=prompt,
-        version=version,
-        ptex=ptex,
-    )
-    info(f"New virtual environment successfully created at {venv_dir}.")
+    if "dev" in str(version):
+        #pikesquares_localdev_dir = os.environ.get("PIKESQUARES_LOCALDEV_DIR")
+        pikesquares_localdev_dir = "/home/pk/dev/eqb/pikesquares"
+        if Path(pikesquares_localdev_dir).exists():
+            debug(f"lift.bindings.install: {pikesquares_localdev_dir=}")
+            install_pikesquares_localdev(
+                venv_dir=venv_dir,
+                pikesquares_localdev_dir=pikesquares_localdev_dir
+            )
+        else:
+            info(f"lift.bindings.install: {pikesquares_localdev_dir=}")
+            sys.exit()
+    else:
+        prompt = f"PikeSquares {version}"
+        install_pikesquares_from_pex(
+            venv_dir=venv_dir,
+            prompt=prompt,
+            version=version,
+            ptex=ptex,
+        )
+        pikesquares_requirements = [f"pikesquares=={version}"]
+        info(
+            f"Installing {' '.join(pikesquares_requirements)} into a virtual environment at {venv_dir}"
+        )
+
+    info(f"New virtual environment successfully created at {venv_dir}")
 
     pikesquares_server_exe = str(venv_dir / "bin" / "pikesquares")
     current_user: str = pwd.getpwuid(os.getuid()).pw_name
@@ -169,7 +201,6 @@ def main() -> NoReturn:
             },
             Query().version == str(version),
         )
-        print(f"wrote venv dir to config. {version=}")
 
     sys.exit(0)